Consolidate decorators
