+++ src/graph/index.js
@@ -1,3 +1,2 @@
-export { default as binding } from './binding';
 export { default as getInjector } from './getInjector';
-export { defaults } from './defaults';
+export { binding, defaults } from './decorators';
